DM_19 small bug fixes
